+++ src/test/kotlin/com/knu/mockin/controller/LoginControllerTest.kt
@@ -10,8 +10,8 @@
     import com.knu.mockin.repository.UserRepository
     import com.knu.mockin.security.JwtUtil
     import com.knu.mockin.security.SecurityTestConfig
-    import com.knu.mockin.service.EmailService
-    import com.knu.mockin.service.UserService
+    import com.knu.mockin.service.login.EmailService
+    import com.knu.mockin.service.login.UserService
     import com.ninjasquad.springmockk.MockkBean
     import io.kotest.core.spec.style.StringSpec
     import io.kotest.extensions.spring.SpringExtension
